main: cast off_t to size_t for cred_ll_readdir pointer arithmetic
- Perform explicit type casting from signed off_t to size_t on FUSE readdir offset when index-accessing the directory response buffer.
- This prevents signed pointer arithmetic compiler warnings and guarantees type-safe pointer addition.

+++ src/main.c
@@ -584,7 +584,7 @@ static void cred_ll_readdir(fuse_req_t req, fuse_ino_t ino, size_t size, off_t o
         if (chunk > size) {
             chunk = size;
         }
-        fuse_reply_buf(req, b.p + off, chunk);
+        fuse_reply_buf(req, b.p + (size_t)off, chunk);
     } else {
         fuse_reply_buf(req, NULL, 0);
     }
